Move DateRange from util.date to lib.date
There is no point in having two separate date utility libraries..
